# Conflicts:
# A lot.

+++ src/community/backup-restore/core/src/test/java/org/geoserver/backuprestore/ExtraFileHandler.java
@@ -4,6 +4,11 @@
  */
 package org.geoserver.backuprestore;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.InputStream;
+import java.io.OutputStream;
 import org.geoserver.backuprestore.tasklet.GenericTaskletHandler;
 import org.geoserver.backuprestore.tasklet.GenericTaskletUtils;
 import org.geoserver.config.GeoServerDataDirectory;
@@ -14,15 +19,7 @@ import org.springframework.batch.core.StepExecution;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.repeat.RepeatStatus;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.InputStream;
-import java.io.OutputStream;
-
-/**
- * Test generic handler that backup and restore an extra file that is not used by GeoServer.
- */
+/** Test generic handler that backup and restore an extra file that is not used by GeoServer. */
 public final class ExtraFileHandler implements GenericTaskletHandler {
 
     public static final String EXTRA_FILE_NAME = "extra_file.properties";
@@ -39,8 +36,11 @@ public final class ExtraFileHandler implements GenericTaskletHandler {
     }
 
     @Override
-    public RepeatStatus handle(StepContribution contribution, ChunkContext chunkContext,
-                               JobExecution jobExecution, BackupRestoreItem context) {
+    public RepeatStatus handle(
+            StepContribution contribution,
+            ChunkContext chunkContext,
+            JobExecution jobExecution,
+            BackupRestoreItem context) {
         File inputDirectory;
         File outputDirectory;
         if (GenericTaskletUtils.isBackup(context)) {
@@ -56,24 +56,30 @@ public final class ExtraFileHandler implements GenericTaskletHandler {
         return RepeatStatus.FINISHED;
     }
 
-    /**
-     * Helper method for copying a file from a directory to another.
-     */
-    private void copyFile(File inputDirectory, String inputFileName, File outputDirectory, String outputFileName) {
+    /** Helper method for copying a file from a directory to another. */
+    private void copyFile(
+            File inputDirectory,
+            String inputFileName,
+            File outputDirectory,
+            String outputFileName) {
         File inputFile = new File(inputDirectory, inputFileName);
         if (!inputFile.exists()) {
             // nothing to copy
             return;
         }
-        File outputFile = new File(outputDirectory, outputFileName);
-        try (InputStream input = new FileInputStream(inputFile);
-             // copy the file to is destination
-             OutputStream output = new FileOutputStream(outputFile)) {
-            IOUtils.copy(input, output);
-        } catch (Exception exception) {
-            throw new RuntimeException(String.format(
-                    "Error copying file '%s' to file '%s'.",
-                    inputFile, outputFile.getAbsolutePath()), exception);
+        if (outputDirectory.exists() && outputDirectory.isDirectory()) {
+            File outputFile = new File(outputDirectory, outputFileName);
+            try (InputStream input = new FileInputStream(inputFile);
+                    // copy the file to is destination
+                    OutputStream output = new FileOutputStream(outputFile)) {
+                IOUtils.copy(input, output);
+            } catch (Exception exception) {
+                throw new RuntimeException(
+                        String.format(
+                                "Error copying file '%s' to file '%s'.",
+                                inputFile, outputFile.getAbsolutePath()),
+                        exception);
+            }
         }
     }
 }
